# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI ポッドキャスト作成・配信プラットフォーム API",
        "title": "Anycast API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/channels/{channelId}/episodes/{episodeId}/script/lines": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "指定したエピソードの台本行一覧を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "script"
                ],
                "summary": "台本行一覧取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "チャンネル ID",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "エピソード ID",
                        "name": "episodeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ScriptLineListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "指定した位置に新しい台本行を追加します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "script"
                ],
                "summary": "台本行作成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "チャンネル ID",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "エピソード ID",
                        "name": "episodeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "台本行作成リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateScriptLineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.ScriptLineResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/channels/{channelId}/episodes/{episodeId}/script/lines/{lineId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "指定した台本行を削除します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "script"
                ],
                "summary": "台本行削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "チャンネル ID",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "エピソード ID",
                        "name": "episodeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "台本行 ID",
                        "name": "lineId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "指定した台本行を更新します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "script"
                ],
                "summary": "台本行更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "チャンネル ID",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "エピソード ID",
                        "name": "episodeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "台本行 ID",
                        "name": "lineId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "台本行更新リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateScriptLineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ScriptLineResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/channels/{channelId}/episodes/{episodeId}/script/reorder": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "台本行の順序を並び替えます",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "script"
                ],
                "summary": "台本行並び替え",
                "parameters": [
                    {
                        "type": "string",
                        "description": "チャンネル ID",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "エピソード ID",
                        "name": "episodeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "並び替えリクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ReorderScriptLinesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ScriptLineListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateScriptLineRequest": {
            "type": "object",
            "required": [
                "speakerId"
            ],
            "properties": {
                "afterLineId": {
                    "type": "string"
                },
                "emotion": {
                    "type": "string"
                },
                "speakerId": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "request.ReorderScriptLinesRequest": {
            "type": "object",
            "required": [
                "lineIds"
            ],
            "properties": {
                "lineIds": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "request.UpdateScriptLineRequest": {
            "type": "object",
            "properties": {
                "emotion": {
                    "type": "string"
                },
                "speakerId": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "response.CharacterVoiceResponse": {
            "type": "object",
            "required": [
                "gender",
                "id",
                "name",
                "provider"
            ],
            "properties": {
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorDetail": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {},
                "message": {
                    "type": "string",
                    "example": "Invalid request parameters"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        },
...
        "response.ScriptLineListResponse": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ScriptLineResponse"
                    }
                }
            }
        },
        "response.ScriptLineResponse": {
            "type": "object",
            "required": [
                "createdAt",
                "id",
                "lineOrder",
                "speaker",
                "text",
                "updatedAt"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "emotion": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lineOrder": {
                    "type": "integer"
                },
                "speaker": {
                    "$ref": "#/definitions/response.SpeakerResponse"
                },
                "text": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "response.SpeakerResponse": {
            "type": "object",
            "required": [
                "id",
                "name",
                "persona",
                "voice"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "persona": {
                    "type": "string"
                },
                "voice": {
                    "$ref": "#/definitions/response.CharacterVoiceResponse"
                }
            }
        },
...
    }
}